+++ src/simulator/domain.py
@@ -199,7 +199,7 @@ class ScalarDomain(eqx.Module):
         del dims
         del valid_types
 
-        predicted_domain_allocation = domain_estimate(self.dims)
+        predicted_domain_allocation = np.int64(domain_estimate(self.dims))
         print("Predicted size in memory of domain:", mem_conversion(predicted_domain_allocation))
 
         if iteration == 1 and auto_batching:
